+++ packages/perseus/src/widgets/interactive-graphs/graphs/index.ts
@@ -1,1 +1,3 @@
 export {SegmentGraph} from "./segment";
+export {LinearGraph} from "./linear";
+export {RayGraph} from "./ray";