+++ packages/perseus/src/widgets/interactive-graphs/graphs/types.ts
@@ -1,0 +1,10 @@
+import type {CollinearTuple} from "../../../perseus-types";
+import type {Interval, vec} from "mafs";
+
+export interface InteractiveLineProps {
+    collinearPair: Readonly<CollinearTuple>;
+    snaps: [number, number];
+    range: [x: Interval, y: Interval];
+    onMovePoint: (endpointIndex: number, destination: vec.Vector2) => unknown;
+    onMoveLine: (delta: vec.Vector2) => unknown;
+}