+++ packages/perseus/src/widgets/interactive-graphs/types.ts
@@ -1,7 +1,13 @@
 import type {InteractiveGraphAction} from "./interactive-graph-action";
-import type {InteractiveGraphState} from "./interactive-graph-state";
-import type {PerseusInteractiveGraphWidgetOptions} from "../../perseus-types";
+import type {
+    CollinearTuple,
+    PerseusGraphTypeLinearSystem,
+    PerseusGraphTypeRay,
+    PerseusGraphTypeSegment,
+    PerseusInteractiveGraphWidgetOptions,
+} from "../../perseus-types";
 import type {WidgetProps} from "../../types";
+import type {Interval, vec} from "mafs";
 
 export type InteractiveGraphProps = WidgetProps<
     PerseusInteractiveGraphWidgetOptions,
@@ -12,3 +18,29 @@ export type MafsGraphProps<T extends InteractiveGraphState> = {
     graphState: T;
     dispatch: (action: InteractiveGraphAction) => unknown;
 };
+
+export type InteractiveGraphState =
+    | SegmentGraphState
+    | LinearGraphState
+    | RayGraphState;
+
+export interface InteractiveGraphStateCommon {
+    hasBeenInteractedWith: boolean;
+    // range = [[xMin, xMax], [yMin, yMax]] in Cartesian units
+    range: [Interval, Interval];
+    // snapStep = [xStep, yStep] in Cartesian units
+    snapStep: vec.Vector2;
+}
+
+export type SegmentGraphState = InteractiveGraphStateCommon &
+    Pick<PerseusGraphTypeSegment, "type" | "coords">;
+
+export type LinearGraphState = InteractiveGraphStateCommon &
+    Pick<PerseusGraphTypeLinearSystem, "coords"> & {
+        type: "linear" | "linear-system";
+    };
+
+export type RayGraphState = InteractiveGraphStateCommon &
+    Pick<PerseusGraphTypeRay, "type"> & {
+        coords: readonly CollinearTuple[];
+    };